+++ xapian-core/common/compression_stream.h
@@ -62,7 +62,7 @@ class CompressionStream {
     void decompress_start() { lazy_alloc_inflate_zstream(); }
 
     /** Returns true if this was the final chunk. */
-    bool decompress_chunk(const char* p, int len, std::string& buf);
+    bool decompress_chunk(const char* p, size_t len, std::string& buf);
 };
 
 #endif // XAPIAN_INCLUDED_COMPRESSION_STREAM_H
